move enrich graph with asset data functionaly to a separate file
