--- x-pack/solutions/security/plugins/cloud_security_posture/server/routes/graph/v1.test.ts
@@ -1,116 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { loggingSystemMock } from '@kbn/core-logging-server-mocks';
-import { getGraph, type GetGraphParams } from './v1';
-import { fetchGraph } from './fetch_graph';
-import { parseRecords } from './parse_records';
-import { SECURITY_SOLUTION_ENABLE_ASSET_INVENTORY_SETTING } from '@kbn/management-settings-ids';
-import { fetchEntityData } from './fetch_entity_data';
-
-jest.mock('./fetch_graph');
-jest.mock('./parse_records');
-jest.mock('./fetch_entity_data');
-
-const mockLoggerFactory = loggingSystemMock.create();
-const mockLogger = mockLoggerFactory.get('mock logger');
-
-describe('getGraph', () => {
-  let esClient: any;
-  let uiSettings: any;
-
-  beforeEach(() => {
-    jest.clearAllMocks();
-    esClient = {};
-    uiSettings = {
-      client: {
-        get: jest.fn(),
-      },
-    };
-
-    // Default to asset inventory disabled
-    uiSettings.client.get.mockResolvedValue(false);
-
-    // Mock fetchEntityData as a no-op since we're not testing that part
-    (fetchEntityData as jest.Mock).mockResolvedValue({});
-  });
-
-  it('should call fetchGraph and parseRecords with correct parameters', async () => {
-    const fakeFetchResult = { records: ['record1', 'record2'] };
-    (fetchGraph as jest.Mock).mockResolvedValue(fakeFetchResult);
-
-    const parsedResult = { nodes: ['node1'], edges: ['edge1'], messages: ['msg1'] };
-    (parseRecords as jest.Mock).mockReturnValue(parsedResult);
-
-    const params = {
-      services: { esClient, logger: mockLogger, uiSettings },
-      query: {
-        originEventIds: [
-          { id: 'event1', isAlert: false },
-          { id: 'event2', isAlert: false },
-        ],
-        spaceId: 'testSpace',
-        indexPatterns: ['pattern1', 'pattern2'],
-        start: 0,
-        end: 100,
-        esQuery: { bool: { must: [{ match_phrase: { field: 'value' } }] } },
-      },
-      showUnknownTarget: true,
-      nodesLimit: 10,
-    } satisfies GetGraphParams;
-
-    const result = await getGraph(params);
-
-    expect(fetchGraph).toHaveBeenCalledWith({
-      esClient,
-      showUnknownTarget: true,
-      logger: mockLogger,
-      start: 0,
-      end: 100,
-      originEventIds: [
-        { id: 'event1', isAlert: false },
-        { id: 'event2', isAlert: false },
-      ],
-      indexPatterns: ['pattern1', 'pattern2'],
-      esQuery: { bool: { must: [{ match_phrase: { field: 'value' } }] } },
-    });
-    expect(parseRecords).toHaveBeenCalledWith(mockLogger, fakeFetchResult.records, 10);
-    expect(uiSettings.client.get).toHaveBeenCalledWith(
-      SECURITY_SOLUTION_ENABLE_ASSET_INVENTORY_SETTING
-    );
-    expect(result).toEqual(parsedResult);
-  });
-
-  it('should use default indexPatterns if not provided', async () => {
-    const fakeFetchResult = { records: [] };
-    (fetchGraph as jest.Mock).mockResolvedValue(fakeFetchResult);
-
-    const parsedResult = { nodes: [], edges: [], messages: [] };
-    (parseRecords as jest.Mock).mockReturnValue(parsedResult);
-
-    const params = {
-      services: { esClient, logger: mockLogger, uiSettings },
-      query: {
-        originEventIds: [{ id: 'event3', isAlert: false }],
-        // No indexPatterns provided; spaceId will be used to build default patterns.
-        spaceId: 'defaultSpace',
-        start: '2020-01-01',
-        end: '2020-01-02',
-        esQuery: { bool: { must: [{ match_phrase: { field: 'value' } }] } },
-      },
-      showUnknownTarget: false,
-    } satisfies GetGraphParams;
-
-    await getGraph(params);
-
-    expect(fetchGraph).toHaveBeenCalledWith(
-      expect.objectContaining({
-        indexPatterns: ['logs-*'],
-      })
-    );
-  });
-});
